Fixed #4 Implemented the kv_client_table.hpp
The const prefix of the AbstractPartitionManager* is preventing the Slice() method to be called, so I have removed the prefix. I assume that it is a bug because the Slice() methods in AbstractPartitionManager are not (and cannot be) constant methods.

Removed the GTEST_FAIL() in TestKVClientTable's Add and Get, these tasks should now pass the tests.